Move ConstantOfShape and Range shape inference into the generate submodule
Move the ConstantOfShape and Range operators, plus their tests, from ops.rs
into ops/generate.rs, matching how these operators are grouped in
src/ops.rs.
